server/requirements.txt
Flask==3.1.2
flask-cors==6.0.2
google-genai==1.62.0
mss==10.1.0
pynput==1.8.1
python-dotenv==1.2.1
